+++ netbox_branching/models/changes.py
@@ -17,6 +17,8 @@ from netbox_branching.utilities import (
     diff_for_merge,
     full_clean_with_file_check,
     resolve_objectchange_field_migration,
+    restore_m2m_values,
+    save_deserialized_object,
     update_object,
 )
 
@@ -95,10 +97,9 @@ class ObjectChange(ObjectChange_):
             if isinstance(instance.object, MPTTModel):
                 clear_mptt_fields(instance.object)
                 instance.object.save(using=using, force_insert=True)
-                for accessor_name, object_list in (instance.m2m_data or {}).items():
-                    getattr(instance.object, accessor_name).set(object_list)
+                restore_m2m_values(instance.object, instance.m2m_data)
             else:
-                instance.save(using=using)
+                save_deserialized_object(instance, using=using)
 
         # Modifying an object
         elif self.action == ObjectChangeActionChoices.ACTION_UPDATE:
@@ -183,10 +184,9 @@ class ObjectChange(ObjectChange_):
             if isinstance(instance, MPTTModel):
                 clear_mptt_fields(instance)
                 instance.save(using=using, force_insert=True)
-                for accessor_name, object_list in (deserialized.m2m_data or {}).items():
-                    getattr(instance, accessor_name).set(object_list)
+                restore_m2m_values(instance, deserialized.m2m_data)
             else:
-                deserialized.save(using=using)
+                save_deserialized_object(deserialized, using=using)
 
     undo.alters_data = True